refactor(parent-grounding-instruction-screen): navigation
