--- app/screens/Grounding/index.ts
@@ -1,1 +1,0 @@
-export * from './General';
