+++ app/screens/ParentGrounding/GroundingInstruction/GroundingInstruction.types.ts
@@ -1,6 +1,6 @@
 import { StackScreenProps, StackNavigationProp } from '@react-navigation/stack';
 
-import { MergedStackParams } from '../../../../navigation/stacks/mergedParams';
+import { MergedStackParams } from '../../../navigation/stacks/mergedParams';
 
 type NavigationProps = StackNavigationProp<
   MergedStackParams,
