+++ app/screens/ParentGrounding/index.ts
@@ -1,0 +1,1 @@
+export * from './GroundingInstruction';
