+++ app/screens/index.ts
@@ -1,6 +1,6 @@
 export * from './CharmAlerts';
 export * from './ComingSoon';
-export * from './Grounding';
+export * from './ParentGrounding';
 export * from './Initial';
 export * from './DummyQuests';
 export * from './Garden';
